vendor_system/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
docutils==0.17.1
filelock==3.0.12
idna==3.2
numpy==1.21.0
pandas==1.2.5
pipenv==2021.5.29
platformdirs==2.2.0
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
scipy==1.7.0
six==1.16.0
sqlparse==0.4.1
statistics==1.0.3.5
urllib3==1.26.6
virtualenv==20.7.0
virtualenv-clone==0.5.6
